test: WALLET-1364 — share the content-script test harness instead of copying it
