+++ src/content/index.test.ts
@@ -1,3 +1,4 @@
+import { FakePort, installFakeDom } from './__fixtures';
 import { sdkMethod } from './sdk-method';
 
 jest.mock('webextension-polyfill', () => ({
@@ -20,85 +21,16 @@ jest.mock('@content/bring', () => ({ initBringScript: jest.fn() }));
 
 const ORIGIN = 'https://dapp.example';
 
-type Handler = (e: { data: unknown }) => void;
-
-// A synchronous stand-in for MessagePort: `postMessage` delivers straight into
-// the peer's `onmessage` and its 'message' listeners.
-class FakePort {
-  peer!: FakePort;
-  onmessage: Handler | null = null;
-  listeners: Handler[] = [];
-  closed = false;
-
-  postMessage = (data: unknown) => {
-    if (this.peer.closed) return;
-    this.peer.onmessage?.({ data });
-    this.peer.listeners.slice().forEach(cb => cb({ data }));
-  };
-  addEventListener = (type: string, cb: Handler) => {
-    if (type === 'message') this.listeners.push(cb);
-  };
-  removeEventListener = (type: string, cb: Handler) => {
-    this.listeners = this.listeners.filter(l => l !== cb);
-  };
-  start = () => undefined;
-  close = () => {
-    this.closed = true;
-  };
-}
-
 const loadContentScript = () => {
-  const messageListeners: Handler[] = [];
+  const { channelRef, scriptTags, messageListeners, cleanupListeners, win } =
+    installFakeDom(ORIGIN);
   const posted: {
     data: { type?: string };
     origin: string;
     transfer: FakePort[];
   }[] = [];
-  const scriptTags: { onload?: () => void; setAttribute: jest.Mock }[] = [];
-  let channel!: { port1: FakePort; port2: FakePort };
-
-  (global as { MessageChannel?: unknown }).MessageChannel = function () {
-    const port1 = new FakePort();
-    const port2 = new FakePort();
-    port1.peer = port2;
-    port2.peer = port1;
-    channel = { port1, port2 };
-    return channel;
-  };
-
-  const cleanupListeners: (() => void)[] = [];
-  const win = {
-    location: { origin: ORIGIN },
-    addEventListener: (type: string, cb: Handler) => {
-      if (type === 'message') messageListeners.push(cb);
-      else cleanupListeners.push(cb as unknown as () => void);
-    },
-    removeEventListener: () => undefined,
-    dispatchEvent: () => true,
-    postMessage: (data: unknown, origin: string, transfer: FakePort[]) => {
-      posted.push({ data: data as { type?: string }, origin, transfer });
-    }
-  };
-  const doc = {
-    head: {
-      children: [{}],
-      insertBefore: () => undefined,
-      removeChild: () => undefined
-    },
-    createElement: () => {
-      const tag = { setAttribute: jest.fn() } as (typeof scriptTags)[number];
-      scriptTags.push(tag);
-      return tag;
-    }
-  };
-
-  (global as { window?: unknown }).window = win;
-  (global as { document?: unknown }).document = doc;
-  (global as { CustomEvent?: unknown }).CustomEvent = class {
-    constructor(
-      public type: string,
-      public init?: unknown
-    ) {}
+  win.postMessage = (data: unknown, origin: string, transfer: FakePort[]) => {
+    posted.push({ data: data as { type?: string }, origin, transfer });
   };
 
   jest.resetModules();
@@ -114,7 +46,14 @@ const loadContentScript = () => {
   // `establishSdkPort` runs from the injected script's onload handler.
   scriptTags[0].onload?.();
 
-  return { channel, posted, messageListeners, cleanupListeners, win, runtime };
+  return {
+    channel: channelRef.current!,
+    posted,
+    messageListeners,
+    cleanupListeners,
+    win,
+    runtime
+  };
 };
 
 beforeEach(() => {
